Refactor: Started to move non-scanning functions out of scanner
